refactor(server): migrate HTTP handlers, auth middleware, and state to inbound server module
