+++ src/server/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod issue_credential;
+pub mod status_list;
+
+pub use issue_credential::credential_handler;
+pub use status_list::*;
