+++ src/server/handlers/status_list/conditional.rs
@@ -1,0 +1,123 @@
+use time::macros::format_description;
+use tracing::warn;
+
+const IMF_FIXDATE: &[time::format_description::BorrowedFormatItem<'static>] = format_description!(
+    "[weekday repr:short], [day] [month repr:short] [year] [hour]:[minute]:[second] GMT"
+);
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ConditionalResponse {
+    NotModified,
+    Modified,
+}
+
+pub fn evaluate_if_none_match(
+    if_none_match: Option<&str>,
+    current_etag: &str,
+) -> ConditionalResponse {
+    let Some(header_value) = if_none_match else {
+        return ConditionalResponse::Modified;
+    };
+
+    let trimmed = header_value.trim();
+    if trimmed == "*" {
+        return ConditionalResponse::NotModified;
+    }
+
+    for etag in header_value.split(',') {
+        let etag = etag.trim();
+        if etag.is_empty() {
+            continue;
+        }
+        if etag_eq_weak(etag, current_etag) {
+            return ConditionalResponse::NotModified;
+        }
+    }
+
+    ConditionalResponse::Modified
+}
+
+pub fn evaluate_if_modified_since(
+    if_modified_since: Option<&str>,
+    updated_at: i64,
+) -> ConditionalResponse {
+    let Some(header_value) = if_modified_since else {
+        return ConditionalResponse::Modified;
+    };
+
+    let Some(client_timestamp) = parse_http_date(header_value) else {
+        warn!("Malformed If-Modified-Since header: {}", header_value);
+        return ConditionalResponse::Modified;
+    };
+
+    let now = time::OffsetDateTime::now_utc().unix_timestamp();
+    if client_timestamp > now {
+        warn!("If-Modified-Since contains future date: {}", header_value);
+        return ConditionalResponse::Modified;
+    }
+
+    if updated_at <= client_timestamp {
+        ConditionalResponse::NotModified
+    } else {
+        ConditionalResponse::Modified
+    }
+}
+
+pub fn evaluate_conditional_request(
+    if_none_match: Option<&str>,
+    if_modified_since: Option<&str>,
+    current_etag: &str,
+    updated_at: i64,
+) -> ConditionalResponse {
+    if if_none_match.is_some() {
+        return evaluate_if_none_match(if_none_match, current_etag);
+    }
+    evaluate_if_modified_since(if_modified_since, updated_at)
+}
+
+pub fn format_http_date(unix_timestamp: i64) -> String {
+    use time::OffsetDateTime;
+
+    let datetime =
+        OffsetDateTime::from_unix_timestamp(unix_timestamp).unwrap_or(OffsetDateTime::UNIX_EPOCH);
+
+    datetime
+        .format(IMF_FIXDATE)
+        .unwrap_or_else(|_| "Thu, 01 Jan 1970 00:00:00 GMT".to_string())
+}
+
+pub fn parse_http_date(date_str: &str) -> Option<i64> {
+    use time::OffsetDateTime;
+
+    OffsetDateTime::parse(date_str, IMF_FIXDATE)
+        .or_else(|_| {
+            OffsetDateTime::parse(date_str, &time::format_description::well_known::Rfc2822)
+        })
+        .ok()
+        .map(|dt| dt.unix_timestamp())
+}
+
+fn normalize_etag(etag: &str) -> String {
+    etag.trim()
+        .trim_start_matches("W/")
+        .trim_matches('"')
+        .to_string()
+}
+
+fn etag_eq_weak(a: &str, b: &str) -> bool {
+    normalize_etag(a) == normalize_etag(b)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_evaluate_if_none_match_single_etag_match() {
+        let current_etag = r#"W/"abc123""#;
+        let if_none_match = r#"W/"abc123""#;
+
+        let result = evaluate_if_none_match(Some(if_none_match), current_etag);
+        assert_eq!(result, ConditionalResponse::NotModified);
+    }
+}
