--- src/web/handlers.rs
@@ -1,8 +1,0 @@
-pub mod issue_credential;
-pub mod status_list;
-
-pub use issue_credential::{CredentialError, credential_handler};
-pub use status_list::{
-    aggregation::get_aggregation, error::StatusListError, get_status_list::get_status_list,
-    publish_status::publish_status, update_status::update_status,
-};
