--- src/web/mod.rs
@@ -1,4 +1,0 @@
-pub mod auth;
-pub mod errors;
-pub mod handlers;
-pub mod rate_limit;
